Fix type and interaction lookup
Summary:
It doesn't make sense to have separate maps for different kinds of top-level definitions because all of them share the same namespace (in the abstract sense, not a Thrift namespace). Now that we properly report redefinitions fix the lookup for types and interactions with the rest to follow.

Apart from better representing reality this will simplify return type handling where we need to distinguish between types and interactions and can improve diagnostics. Instead of incorrectly reporting that a name is undefined we can report that it is defined as a wrong kind of entity.

Reviewed By: iahs

Differential Revision: D48832714

fbshipit-source-id: 44acbac5852d71a2ae98634f7a17b463224b26da