t_go_generator: access has_doc() only via t_named
Summary:
To facilitate upcoming move of doc from t_node to t_named. While working on this code, to help me better understand the flow and improve encapsulation, I moved the relevant methods to the private section and renamed them to clarify which of the overloaded methods was actually called when. If this is an issue, I can revert these specific changes.

Edit: reverted method rename (per vitaut's comment in stacked diff)

Reviewed By: echistyakov

Differential Revision: D48849722

fbshipit-source-id: 46cc44a9b8a2ae27de53da50a393aaf2f940759e