(rsocket java) upgrade rsocket version
Summary:
rsocket java version is lagging behind (current version 1.0.4 was release in 2021). In this diff, we upgrade rsocket java version to 1.1.4 which include several buffer leak fixes.

Some rsocket-java internal API change makes `ReactorDuplexConnection` no longer compatible with 1.1.4.  The original reason why we need `ReactorDuplexConnection` is that we would like to add `alpn` support there. Now that we would like to use server proxy to deal with protocol negotiation, we no longer need this class. We migrate to use `TcpDuplexConnection` provided by `rsocket-transport-netty` instead.

Note that on rsocket-java 1.1.4, `rsocket.onClose()` is no longer a `MonoProcessor`. so we can no longer cast return of `rsocket.onClose()` to `MonoProcessor`. We dealt with this issue with D48939678.

Reviewed By: robertroeser

Differential Revision: D48922182

fbshipit-source-id: 62550514745c6cbe51f226d0899e2900096b1183