Revert D48763403: Multisect successfully blamed "D48763403: [thrift] Add client_methods_bypass_eventhandlers Thrift Flag" for test or build failures
Summary:
This diff is reverting D48763403
D48763403: [thrift] Add client_methods_bypass_eventhandlers Thrift Flag by praihan has been identified to be causing the following test or build failures:

Tests affected:
- [mandoline/alertrouter/tests:alert_handler_test - HandlerTest.TestOriginalClientIdPropagation](https://www.internalfb.com/intern/test/281475051596075/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2995727
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: jdonald

Differential Revision: D49114117

fbshipit-source-id: 2198ba38644ad6324642b44b481804f5e2a63610