Revert D48724874: Multisect successfully blamed "D48724874: [thrift] Populate PreHandlerTimestamps regardless of sampling status" for test or build failures
Summary:
This diff is reverting D48724874
D48724874: [thrift] Populate PreHandlerTimestamps regardless of sampling status by sazonovkirill has been identified to be causing the following test or build failures:

Tests affected:
- [ame/serf/base/tests:base - BaseTestNoDependency.testSetDeviceSignal](https://www.internalfb.com/intern/test/281475069084418/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3020674
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: davidrothera

Differential Revision: D49219147

fbshipit-source-id: 3c3ff521acd00ee84357c13019bfe7386b77b659