Accurately reflect paths from IDL in AST
Summary:
The AST is meant to be an accurate representation of the contents of a thrift file. The interaction with the file system, notably around searching include paths to resolve includes, should be encapsulated in the source manager and not leak into the returned AST.
The resolved path is stored in source_manager so get_file continues to return the right content.

Reviewed By: thedavekwon

Differential Revision: D49429380

fbshipit-source-id: 9b5792a43d28457273c60d47f11b0dec6b89f1ff