cython helper for replacing thrift-py3 private api usage
Summary:
Thrift-py3 power users use the "private py3 api" for zero-cost FFI. The main ones are `py3struct._cpp_obj` and `Py3Struct._fbthrift_create`. These block migration from thrift-py3 to thrift-python.

This diff provides a cython wrapper around the python_capi `Constructor` and `Extractor`. The benefit of this approach is that we do the error handling safely. Unfortunately, this makes the user responsible for instantiating the templated cython function by binding the C++ thrift type. The user is also responsible for including the python_capi dependencies.

This is unfortunate, but folks who aren't going to use ligen should expect to do some extra work.

`lib.pyx` in the test is example of what this looks like, mixed together with binding some random C++ code.

If this is approved, I'll add a how-to section to
https://www.internalfb.com/intern/wiki/Thrift-py3/User_Guide/Advanced_Usage/Converting_from_and_to_a_Thrift-cpp2_struct/

And add a link to:
https://www.internalfb.com/intern/wiki/Thrift-Python/Migration_Guide/Pitfalls_and_Tips/From_thrift-py3/

Reviewed By: Filip-F

Differential Revision: D49887238

fbshipit-source-id: e9d6292c745556c68b8f74110bea553a4df1e689