Fix exceptions for C++ clients
Summary:
We were writing blank "Success" value on the wire in case of an exception.
This was confusing C++ clients into thinking that there was no exception thrown and a legit response was returned instead. Even though exception was present in the response.

To be compatible with C++ Thrift implementation - we will now mark "Success" response as optional and nothing will be written to the wire when an exception is encountered/returned.

This is also consistent with the legacy generator (which uses optional qualifier for "Success" return value).

Reviewed By: awalterschulze

Differential Revision: D50024477

fbshipit-source-id: f85f9471bc24b0500aa8a5e764e05d1ec62809a7