mark requests as APP_OVERLOADED when the remote end sends a 503 with Retry-After
Summary:
HTTP endpoints may return 503 with a `Retry-After` header set when there's downtime, or when there is no valid upstream behind a reverse proxy; the latter is also desirable to return in overload situations either from the reverse proxy, or from the HTTP server if requests are dropped from queues.

Here, we merely look at the presence of the header, and do not parse and obey the values… but it's either a `http-date` or a number of seconds.

We do not implement this for 429 currently, since that is meant for quota violations, and would require actually parsing and obeying the value.

Note that this is not a connect error, since semantically the connection to the remote end is perfectly valid and a HTTP response. It just wants us to fuck off for a bit.

Differential Revision: D49954324

fbshipit-source-id: 80d768608473fadd017e6930df615e562558befd