tuple index based on field_ordered_by_key
Summary:
Got UB when enabling marshal capi on thrift structs with out-of-order keys. That's because thrift-python internal tuple stores fields as sorted by key, not the "index" (declaration order in thrift file).

This adds logic to lookup the proper tuple index based on key-sort order. It's a bit tricky to do this; the `mstch_field` doesn't have enough info to evaluate, so evaluate in `mstch_struct` and put into an array where the value at each index (declaration order) is the thrift-python dataTuple index position.

Reviewed By: Filip-F

Differential Revision: D50356322

fbshipit-source-id: 508447468a814400d8a180419ca166d168b32709