Enable user provided attributes on service
Summary:
Internal :
OneCatalog parse the hack files and associates an owner to each class based on the <<Oncalls>> attribute. Thrift generated file don't have this attribute on each generated class as at the time of codegen we don't have the owner information. This attribute is anyways meta specific so even if the info was available, we cannot generate this attribute via codegen. We instead add oncall information at the top of the file in the generated header which is added to the files as part of meerkat step.

Enabling `hack.Atributes` on service allows users to inject `<<Oncall>>` attribute on generated service/client classes.

More details here -  https://fburl.com/workplace/ovp78uaj

(Note: this ignores all push blocking failures!)

Reviewed By: thedavekwon

Differential Revision: D50394731

fbshipit-source-id: 882e94b8c5cbf88a0f193eb907b7a1ba9f13e8b6