Attempt to fix duplicate flag issue (#579)
Summary:
X-link: https://github.com/facebook/hhvm/pull/9400

Pull Request resolved: https://github.com/facebook/fbthrift/pull/579

CacheLib noticed OSS build failures due to `JSONProtocolCommon.cpp` being linked into the build twice, after we added a GFLAG to that file.

This diff attempts to fix the issue by depending on `thriftprotocol` rather than duplicating the files.

The code claims that it couldn't depend on `thriftprotocol` due to a circular dependency (added in D24720566), but Sandcastle and GitHub Actions is passing, so it doesn't seem to still be an issue.

Reviewed By: vitaut

Differential Revision: D50505357

fbshipit-source-id: 785ee83ea9d36514599ee7dd5c69095e7b6e93ec