Fix duplicate flags in OSS build (take 2) (#580)
Summary:
Pull Request resolved: https://github.com/facebook/fbthrift/pull/580

D50505357 attempted to fix the issue, but introduced dependency cycles that weren't caught by CI.

This diff fixes the issue by moving all the GFlag definitions to `ProtocolGFlags.cpp` and exposing this as a separate library `thriftprotocolflags` in CMake. Each library which includes a copy of the protocol sources depends on this library.

Differential Revision: D50521352

fbshipit-source-id: f27a7ce8ff5e75a87dc9397e448c011a980b5984