Remove gflag in JSONProtocolCommon
Summary: This is an alternative diff to D50521352, that simply removes the gflag causing the [issue](https://github.com/facebook/CacheLib/issues/267). This is likely an inferior option, since it removes the option for users to opt ouf of the new behavior of ignoring padding in base64.  However, while this was technically a breaking change, since previously the base64 decoder would append garbage bytes to the end of the input in the presence of padding, it seems unlikely that anyone depended on this corruption.

Reviewed By: vitaut, jaesoo-fb

Differential Revision: D50568555

fbshipit-source-id: bdebd070bd1d620c653d619317e9898eac049ca5