Fix IG wheel: forward/backward compat for codegen thrift primitives
Summary:
Even though D50523077 passed `igenv-test`, it appears to have broken the wheel land-time job `async-job-compatibility-check`. The issue is that even though most IG wheel pipeline jobs regenerate thrift before doing anything, this job does not, so the new logic gets a tuple index failure due to the missing field in the old generated python.

https://www.internalfb.com/sandcastle/workflow/954763121003937298

In my opinion this is an issue with the above job that could be fixed by not depending on committed generated code, or by running the igthrift regenerate script like everything else in the wheel build pipeline. However, this diff is another alternative for unblocking the wheel.

This is a forward/backward compatibility fix that leaves the new codegen'd "is primitive" field while using the old `types.pyx` logic. Once all the igthrift is updated, then we can re-land the `types.pyx` logic that relies on presence of the "is primitive" field

Reviewed By: Filip-F

Differential Revision: D51077730

fbshipit-source-id: dfb5e8e7a207e4724d193e7f3f3b0ef7c1f3e557