Disallow null values in marked-as-set union fields
Summary:
If a thrift union is initialized with a null value in JAVA, ex:

  union UnionField {
   1: string stringValue;
  }

`UnionField v = UnionField.fromStringValue(null);`

During serialization, the null field will be ignored and the union may be serialized to an empty byte array. **This invalid behavior means that serializing such union succeeds without error, while the resulting data cannot be deserialized**.

After discussion, it was decided that such Union (with fields marked as set but having null values) is an invalid state. (e.g. D28605409 should be reverted).

This change effectively reverts D28605409 and takes a further step of disallowing Union to be initialized with null values in the first place.

This is covered with a System property to allow individual services to revert the behavior in the case this causes a severe regression. To override, set:
`-Dthrift.union.allow-construct-from-null=true`

Reviewed By: avalonalex

Differential Revision: D50917033

fbshipit-source-id: 8a845a2ed8ad6aa8dae47d7069d78cc9bb187561