Remove is_interaction_member
Summary:
`is_interaction_member` is not needed because whether a function is a part of an interaction can and should be checked by looking at the interface node that contains the function. In general having such information about parent nodes is discouraged. Now that we have a proper way to query this information in generators, remove `is_interaction_member` and a corresponding setter.

As a nice side effect this replaces an interaction check for every function in automock with one check per service.

Reviewed By: thedavekwon

Differential Revision: D51604686

fbshipit-source-id: 8efdfbc936596b6532e55687a251fe7fc26b47c2