Remove extraneous semicolon from thrift/compiler/generate/t_json_generator.cc
Summary:
Extraneous semicolons are a code smell and can mask more serious problems. `-Wextra-semi-stmt` finds them.

This diff removes an extraneous semicolon or adjusts a macro so that a semicolon is required after the macro (making it look like a standard function).

This file is drawn from a heavy-hitting list, so fixing this problem will allow *many* other files to take advantage of the safety `-Wextra-semi-stmt` offers.

This should be a low-risk diff: if it compiles, it works.

Reviewed By: thedavekwon

Differential Revision: D51631109

fbshipit-source-id: ac3b3c00f762c9046196738794b8d3519ed89cb6