Remove extraneous semicolon from thrift/lib/cpp2/protocol/TableBasedSerializerImpl.h
Summary:
Extraneous semicolons are a code smell and can mask more serious problems. `-Wextra-semi-stmt` finds them.

This diff removes an extraneous semicolon or adjusts a macro so that a semicolon is required after the macro (making it look like a standard function).

This file is drawn from a heavy-hitting list, so fixing this problem will allow *many* other files to take advantage of the safety `-Wextra-semi-stmt` offers.

This should be a low-risk diff: if it compiles, it works.

Reviewed By: thedavekwon

Differential Revision: D51631125

fbshipit-source-id: d68bd604c7b5cf5a41f6dadd6a41a227e72bb8cb