Remove extraneous semicolon from thrift/compiler/generate/t_java_deprecated_generator.cc
Summary:
Extraneous semicolons are a code smell and can mask more serious problems. `-Wextra-semi-stmt` finds them.

This diff removes an extraneous semicolon or adjusts a macro so that a semicolon is required after the macro (making it look like a standard function).

This file is drawn from a heavy-hitting list, so fixing this problem will allow *many* other files to take advantage of the safety `-Wextra-semi-stmt` offers.

This should be a low-risk diff: if it compiles, it works.

Reviewed By: meyering, vitaut

Differential Revision: D51631146

fbshipit-source-id: 1ded7584955950418f5149ccc6245a74c490aaaf