Add ServiceModule with scoped TProcessorEventHandler support
Summary:
This diff introduces the concept of a (named) `ServerModule`. Every module must have a unique name, primarily for debugging purposes. This restriction can be relaxed in the future if necessary.

Currently a `ServerModule` only encapsulates a set of `TProcessorEventHandler`s but this will change in the future.

`TProcessorEventHandler` will be eventually replaced by an alternative, more robust, interception method — hence they are referred to as "legacy" event handlers in this diff.

At present, `TProcessorEventHandler` instances are global (added via `TProcessorBase::addProcessorEventHandler`). This means that all `ThriftServer` instances share the same set of event handlers which can be problematic:
1. We cannot have a specific server-scoped event handler. Let's say, for example, if we want some extra ACL checks for only one service.
2. `TProcessorEventHandler` implementations must account for storing state for more than one `ThriftServer` instance. The current workaround for this is the `void* ctx` abstraction afforded by `TProcessorEventHandler::getServiceContext` but this is ugly and very error-prone.

After this diff, `TProcessorEventHandler` instances are scoped to each `ThriftServer` instance. As a fallback, before a Thrift server begins it will automatically add all globally registered `TProcessorEventHandler` instances to its scoped set so that existing behavior is preserved. This assumes that no user is adding *global* `TProcessorEventHandler` instances after the server has started running (they can add via `AsyncProcessor::addEventHandler` just fine). A quick code search[1] shows that this assumption is safe.
 ---
In summary...

**Before**:
There are two sources of `TProcessorEventHandler` instances — the *global* set and those added manually to each `AsyncProcessor` instance.

**After**:
There are three sources of `TProcessorEventHandler` (that are coalesced into two) — the *global* set, those added to each `AsyncProcessor`, and those added to `ThriftServer` instances via server modules. `ThriftServer` automatically pulls from the *global* set at startup so that the runtime implementation no longer needs to look at the global set again. This isolates the global set from the rest of the codebase and paves the way for its removal (after migration) in the future.

Reviewed By: maalbash

Differential Revision: D51483690

fbshipit-source-id: 92854493e326d2babded7468b0962764abb78e28