Restructure union destruction
Summary: The dead store to type_ in the destructor is showing up in profiles for some hot code. Previous attempts to fix this by duplicating the field destruction regressed perf likely due to increased code size. Instead we split out the destruction into a new function so it can be shared between clear and the destructor.

Reviewed By: vitaut

Differential Revision: D51820221

fbshipit-source-id: 845df1c5decfa774da275fa627b42ab334bf1be9