remove rust.type hack
Summary:
D51368304 concerned with the implementation of `rust.type` annotations, left behind some "special case" logic for `bytes::Bytes` to exclude the generation of serialization code when encountered as is normally emitted for "nonstandard" types.

it has since come to light (see documentation in D51721781) that the root cause for the necessity of this logic are annotations of the form `rust.type="bytes::Bytes"` in fbcode: this is wrong, `bytes::Bytes` is standard and so the correct form of these annotations is `Bytes`.

this diff fixes the incorrect annotations and removes the hack.

Reviewed By: zertosh, emersonford

Differential Revision: D51823800

fbshipit-source-id: 37d8580d93723a00fcf7ce5da1f60af6005d5dc5