Revert D51789318: Multisect successfully blamed "D51789318: [thrift] Fix issue that diff(...) can not handle AssignOnly patch" for otest failure
Summary:
This diff is reverting D51789318
D51789318: [thrift] Fix issue that diff(...) can not handle AssignOnly patch by Mizuchi has been identified to be causing the following test failure:

Tests affected:
- [icsp/repo/transcoder/test:transcode_service_handler_test - TranscoderTest.ToYamlPatchEnumString](https://www.internalfb.com/intern/test/281475074291228/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3710120
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: milend

Differential Revision: D51939363

fbshipit-source-id: 509462f149f8a84ead6e505957ffa5a6c7cf2829