Revert D51954577: Multisect successfully blamed "D51954577: [thrift] Fix issue that diff(...) can not handle AssignOnly patch (redo)" for otest failure
Summary:
This diff is reverting D51954577
D51954577: [thrift] Fix issue that diff(...) can not handle AssignOnly patch (redo) by Mizuchi has been identified to be causing the following test failure:

Tests affected:
- [fbjava/unified-metastore/um-access:um-access-test-java11 - testPatchAnnotationsInvalidAnnotationPath (com.facebook.di.um.definition.TestAnnotationApis)](https://www.internalfb.com/intern/test/562950030266120/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3746672
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: Adolfo-Karim

Differential Revision: D52061348

fbshipit-source-id: bc621166a23656453446d9beae90f2112f126510