Remove name_index
Summary: `name_index` is a virtually unused wrapper around `std::unordered_map` with a weird, error-prone API. Most confusingly, some of the insertion methods use the node's name and some take a name as a separate argument which suggests that these should be different types of containers (map and set). Replace the few uses of `name_index` with `std::unordered_map` which provides a well-understood semantics. Once we can use unordered containers with heterogeneous lookups we can replace the one in `t_structured` with a set.

Reviewed By: Mizuchi

Differential Revision: D52421603

fbshipit-source-id: 8ef2079b5ac2d4e107a92c59d28c90bc8eab6643