thrift-rust: compilation failure without either deprecated_default_enum_min_i32 or default_enum_zero
Summary:
Thrift enums should have the default value of zero according to the standard. The current behavior is to initialize it to `i32::MIN`, unless opted in to the standard behavior via `default_enum_zero` flag.

This change is a part of the effort to change behavior so the correct one becomes the default one and the legacy non-spec-conforming one is an opt-in.

In order to transition to the new behavior (which is a breaking change), a thrift compiler flag `deprecated_default_enum_min_i32` is introduced. In the first stage, it will be a compilation error to use enums in rust unless wither of the two flags is used. In the following changes, this limitation will be relaxes and the current opt-in `default_enum_zero` behavior will become the default one.

Reviewed By: emersonford

Differential Revision: D52334311

fbshipit-source-id: 617ed91fc26d39f4c6db05666bce0148afcd3078